Include login logic into splash screen. Also split sliding panel into different component
