+++ src/screens/Splash.tsx
@@ -1,20 +1,36 @@
 import { StackScreenProps } from "@react-navigation/stack";
-import React from "react";
-import { Image, StyleSheet, Text, TouchableOpacity, View } from "react-native";
-import SlidingUpPanel from "rn-sliding-up-panel";
+import React, { useEffect } from "react";
+import { StyleSheet, Text, TouchableOpacity, View } from "react-native";
 import { LinearGradient } from "expo-linear-gradient";
+import AsyncStorage from "@react-native-async-storage/async-storage";
 
 import { StackParams } from "../../App";
 
-import sonrLogo from "../../assets/sonr.png";
-import { PrimaryButton } from "../components/PrimaryButton";
 import BeamLogo from "../icons/Beam";
-import { SecondaryButton } from "../components/SecondaryButton";
+import { completeLogin } from "../lib/login";
+import { loginWithSession } from "../lib/matrix";
 
 type Props = StackScreenProps<StackParams, "Login">;
 
 const SplashScreen: React.FC<Props> = ({ navigation }) => {
   let panelRef: any = null;
+
+  const loadSession = async () => {
+    const sessionUser = await AsyncStorage.getItem("sessionUser");
+    const sessionToken = await AsyncStorage.getItem("sessionToken");
+
+    if (!sessionUser || !sessionToken) {
+      navigation.navigate("Login", {});
+      return;
+    }
+
+    const client = await loginWithSession(sessionUser, sessionToken);
+    await completeLogin(client, navigation);
+  };
+
+  useEffect(() => {
+    loadSession();
+  }, []);
 
   return (
     <>
@@ -35,43 +51,6 @@ const SplashScreen: React.FC<Props> = ({ navigation }) => {
           <BeamLogo width="355" height="189" />
           <Text style={styles.h1}>Beam</Text>
         </TouchableOpacity>
-        <SlidingUpPanel
-          ref={(c) => {
-            panelRef = c;
-          }}
-        >
-          <View style={styles.slidePanel}>
-            <LinearGradient
-              style={[styles.gradientContainer]}
-              colors={["#686375", "#1D1A27"]}
-              locations={[0.0, 1]}
-              start={{ x: 0.6, y: 1.5 }}
-              end={{ x: 0.4, y: -0.5 }}
-            />
-            <Text style={styles.subtitle2}>BEAM CONNECTS WITH</Text>
-
-            <Image
-              source={sonrLogo}
-              style={styles.sonrLogo}
-              onLoadEnd={() => panelRef.show(320)}
-            />
-
-            <PrimaryButton
-              onPress={() => navigation.navigate("Login", {})}
-              text="Create Account"
-            />
-
-            <SecondaryButton
-              onPress={() => navigation.navigate("Login", {})}
-              text="Secure Login"
-              style={{ marginTop: 10 }}
-            />
-
-            <Text onPress={() => panelRef.hide()} style={styles.skipText}>
-              Skip
-            </Text>
-          </View>
-        </SlidingUpPanel>
       </View>
     </>
   );
@@ -87,12 +66,6 @@ const styles = StyleSheet.create({
     flex: 1,
     width: "100%",
   },
-  slidePanel: {
-    flex: 1,
-    alignItems: "center",
-    backgroundColor: "#686375",
-    borderRadius: 36,
-  },
   h1: {
     fontFamily: "THICCCBOI_ExtraBold",
     fontSize: 34,
@@ -101,31 +74,6 @@ const styles = StyleSheet.create({
     letterSpacing: -0.01,
     textAlign: "center",
     color: "white",
-  },
-  subtitle2: {
-    fontFamily: "THICCCBOI_Regular",
-    fontSize: 14,
-    fontWeight: "700",
-    lineHeight: 20,
-    letterSpacing: 0.04,
-    textAlign: "left",
-    color: "white",
-    marginTop: 24,
-    marginBottom: 14,
-  },
-  sonrLogo: {
-    height: 62,
-    width: 182,
-    marginBottom: 40,
-  },
-  skipText: {
-    fontFamily: "THICCCBOI_ExtraBold",
-    fontSize: 14,
-    fontWeight: "800",
-    lineHeight: 16,
-    letterSpacing: 0.02,
-    color: "#1792FF",
-    marginTop: 24,
   },
 });
